Rename Room to Zone
This commit renames the Room resource to Zone.

This makes changes both the schema and all the models, views,
controllers, helpers and assets.

This is part of a new naming scheme which abstracts the concepts a
little more so they're not tied to rooms of a building but can be used
to represent any zone that you can check into.

All zones are part of an area.