Add global loading indicator
This commit adds a global loading bar which can be triggered from
anywhere in the app. Basically you just tell it when to start and stop
and it will manage multiple loading requests.

Other changes include a fix to the AJAX zone refreshing. The backbone
views weren't distinct and checking in to one area did not refresh the
other area, that's now fixed.